+++ pyxdameraulevenshtein/pyxdameraulevenshtein.pyx
@@ -23,13 +23,16 @@
 
 from libc.stdlib cimport calloc, free
 
-cdef unicode to_unicode(char *s):
+cdef unicode to_unicode(s):
 	"""
-		Convert char * to a Python unicode object.
+		Convert s to a proper unicode type (handles difference between Python 2 and 3). Code comes from
+		https://groups.google.com/d/msg/cython-users/ofT3fo48ohs/rrf3dtbHkm4J
 	"""
-	return s.decode('UTF-8', 'strict')
+	if isinstance(s, bytes):
+		return (<bytes>s).decode('UTF-8')
+	return s
 
-cpdef unsigned int damerau_levenshtein_distance(char *seq1, char *seq2):
+cpdef unsigned int damerau_levenshtein_distance(seq1, seq2):
 	"""
 		Return the edit distance. This implementation is based on http://mwh.geek.nz/2009/04/26/python-damerau-levenshtein-distance/
 		and runs in O(N*M) time using O(M) space. Because this returns a C unsigned int, the corresponding Python data type will be long (see
@@ -38,11 +41,11 @@ cpdef unsigned int damerau_levenshtein_distance(char *seq1, char *seq2):
 		Examples:
 
 		>>> damerau_levenshtein_distance('smtih', 'smith')
-		1L
+		1
 		>>> damerau_levenshtein_distance('saturday', 'sunday')
-		3L
+		3
 		>>> damerau_levenshtein_distance('orange', 'pumpkin')
-		7L
+		7
 	"""
 
 	s1 = to_unicode(seq1)
@@ -59,49 +62,49 @@ cpdef unsigned int damerau_levenshtein_distance(char *seq1, char *seq2):
 		return len(s1)
 
 	cdef unsigned int i, j, delete_cost, add_cost, subtract_cost, edit_distance
-	cdef unsigned int *two_ago = <unsigned int *>calloc(len(s2) + 1, sizeof(unsigned int))
-	if not two_ago:
-		raise MemoryError()
-	cdef unsigned int *one_ago = <unsigned int *>calloc(len(s2) + 1, sizeof(unsigned int))
-	if not one_ago:
-		raise MemoryError()
-	cdef unsigned int *this_row = <unsigned int *>calloc(len(s2) + 1, sizeof(unsigned int))
-	if not this_row:
+	cdef unsigned int s2_length = len(s2)
+	cdef unsigned int offset = s2_length + 1
+
+	#storage is a 3 x (len(s2) + 1) array that stores two_ago, one_ago, and this_row
+	#multipliers:
+	#two_ago = 0
+	#one_ago = 1
+	#this_row = 2
+	cdef unsigned int *storage = <unsigned int *>calloc(3 * offset, sizeof(unsigned int))
+	if not storage:
 		raise MemoryError()
 
 	try:
 		#initialize this_row
-		for i in xrange(1, len(s2) + 1):
-			this_row[i - 1] = i
+		for i in xrange(1, offset):
+			storage[2 * offset + (i - 1)] = i
 		
 		for i in xrange(len(s1)):
 			#swap/initialize vectors
-			for j in xrange(len(s2) + 1):
-				two_ago[j] = one_ago[j]
-				one_ago[j] = this_row[j]
-			for j in xrange(len(s2)):
-				this_row[j] = 0
-			this_row[len(s2)] = i + 1
+			for j in xrange(offset):
+				storage[j] = storage[offset + j]
+				storage[offset + j] = storage[2 * offset + j]
+			for j in xrange(s2_length):
+				storage[2 * offset + j] = 0
+			storage[2 * offset + s2_length] = i + 1
 
 			#now compute costs
-			for j in xrange(len(s2)):
-				delete_cost = one_ago[j] + 1
-				add_cost = this_row[j - 1 if j > 0 else len(s2)] + 1
-				subtract_cost = one_ago[j - 1 if j > 0 else len(s2)] + (s1[i] != s2[j])
-				this_row[j] = min(delete_cost, add_cost, subtract_cost)
+			for j in xrange(s2_length):
+				delete_cost = storage[offset + j] + 1
+				add_cost = storage[2 * offset + (j - 1 if j > 0 else s2_length)] + 1
+				subtract_cost = storage[offset + (j - 1 if j > 0 else s2_length)] + (s1[i] != s2[j])
+				storage[2 * offset + j] = min(delete_cost, add_cost, subtract_cost)
 				#deal with transpositions
 				if (i > 0 and j > 0 and s1[i] == s2[j - 1] and s1[i - 1] == s2[j] and s1[i] != s2[j]):
-					this_row[j] = min(this_row[j], two_ago[j - 2 if j > 1 else len(s2)] + 1)
-		edit_distance = this_row[len(s2) - 1]
+					storage[2 * offset + j] = min(storage[2 * offset + j], storage[j - 2 if j > 1 else s2_length] + 1)
+		edit_distance = storage[2 * offset + (s2_length - 1)]
 	finally:
-		#be sure to free all dynamically-allocated memory
-		free(two_ago)
-		free(one_ago)
-		free(this_row)
+		#free dynamically-allocated memory
+		free(storage)
 
 	return edit_distance
 
-cpdef double normalized_damerau_levenshtein_distance(char *seq1, char *seq2):
+cpdef double normalized_damerau_levenshtein_distance(seq1, seq2):
 	"""
 		Return a real number between 0.0 and 1.0, indicating the edit distance as a fraction of the longer
 		string. 0.0 means that the sequences are identical, while 1.0 means they have nothing in common.