+++ CHANGES.md
@@ -1,4 +1,8 @@
 # Changes
 
+## 1.1 (2013-09-20)
+* Fixed Python 3 unicode issue (thanks to Stefan Behnel - https://groups.google.com/d/msg/cython-users/ofT3fo48ohs/rrf3dtbHkm4J).
+* Fixed a possible memory leak (thanks to Stefan Behnel).
+
 ## 1.0 (2013-07-03)
 * Initial release.